requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

openai-whisper
torch>=2.0
torchvision
torchaudio
librosa>=0.10
numpy>=1.23
pandas>=1.5
nltk>=3.8
matplotlib>=3.7
soundfile>=0.12
tqdm>=4.65
